pyproject.toml
[project]
name = "whatsapp-commerce"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "alembic>=1.18.4",
    "apscheduler>=3.11.2",
    "asyncpg>=0.31.0",
    "cloudinary>=1.44.2",
    "fastapi[standard]>=0.136.1",
    "httpx>=0.28.1",
    "langchain>=1.2.17",
    "langchain-openai>=1.2.1",
    "langgraph>=1.1.10",
    "loguru>=0.7.3",
    "openai>=2.34.0",
    "psycopg2-binary>=2.9.12",
    "pwdlib[argon2]>=0.3.0",
    "pydantic[email]>=2.13.4",
    "pyjwt>=2.12.1",
    "pymysql>=1.1.3",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.27",
    "pytz>=2026.2",
    "sqlalchemy>=2.0.49",
    "supabase>=2.30.0",
    "tiktoken>=0.12.0",
    "unstructured[all-docs]>=0.22.27",
]
